Adding some simple compression on the pixel data.
Removing the alpha channel in the 32bit pixel cost too much. In most
cases its faster to just send the extra 8 bits.

So, to lower the amount of data I have repurposed the alpha channel to
encode a repeat signal on a given pixel. This allows any congestive
pixels to be removed from the buffer and not sent over the wire.